Upstream version 7.1.11

# gpg: Signature made Sun 01 Feb 2015 09:16:04 PM CET using RSA key ID FD030EEA
# gpg: Good signature from "Matthias Geerdsen <matthias.geerdsen@gmail.com>"
# gpg:                 aka "Matthias Geerdsen <matthias@vorlons.info>"
# gpg:                 aka "Matthias Geerdsen <matthias.geerdsen@intranda.com>"
# gpg:                 aka "Matthias Geerdsen <matthias@geerdsen.net>"
# gpg:                 aka "Matthias Geerdsen <vorlon@vorlons.info>"
# gpg:                 aka "Matthias Geerdsen <mg@geerdsen.net>"
# gpg:                 aka "Matthias Geerdsen <vorlon@fsfe.org>"
# gpg:                 aka "Matthias Geerdsen <mail@geerdsen.net>"

+++ provider/libserver/ECSessionGroup.cpp
@@ -478,7 +478,8 @@ ECRESULT ECSessionGroup::AddChangeNotification(ECSESSIONID ulSessionId, unsigned
 	notification	notifyItem = {0};
 	notificationICS	ics = {0};
 	entryId			syncStateBin = {0};
-	notifySyncState	syncState = {ulSyncId, ulChangeId};
+
+	notifySyncState	syncState = { ulSyncId, static_cast<unsigned int>(ulChangeId) };
 	SESSIONINFOMAP::iterator iterSessions;
 
 	notifyItem.ulEventType = fnevZarafaIcsChange;
